Refactor file names for readability and consistency
Renamed files to improve readability and maintain consistency across the project.
