+++ src/apps/manajemen-piket/handlers.ts
@@ -4,7 +4,7 @@ import {
   updateMember,
   deleteMember,
   readAllMembers,
-} from "../manajemen_piket/manajemen-piket";
+} from "./manajemen-piket";
 import { StatusCodes } from "http-status-codes";
 import { logger } from "../../utils/logger";
