+++ src/server.ts
@@ -5,9 +5,9 @@ import helmet from "helmet";
 import cors from "cors";
 import { logger } from "./utils/logger";
 import { LihatDaftarPiket } from "./apps/lihat_daftar_piket/types";
-import manajemenPiketRoutes from "./apps/manajemen_piket/routes";
-import daftarAdminRoute from "./apps/daftar_admin/daftar-admin-route";
-import loginRoute from "./apps/login/login-route";
+import manajemenPiketRoutes from "./apps/manajemen-piket/routes";
+import daftarAdminRoute from "./apps/daftar-admin/routes";
+import loginRoute from "./apps/login/routes";
 
 dotenv.config();
 const app = express();
